requirements.txt
fastapi
uvicorn
stripe
python-dotenv
pydantic
email-validator
